% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_coverage_for_type_and_chromosome}
\alias{plot_coverage_for_type_and_chromosome}
\title{Plot the distribution of mutation types in the database}
\usage{
plot_coverage_for_type_and_chromosome(endpoint, type, chromosome)
}
\arguments{
\item{endpoint}{The endpoint address to get data from.}

\item{type}{The type to plot the}
}
\description{
Plot the distribution of mutation types in the database
}
