% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lengths_for_type_and_chromosome.R
\name{plot_lengths_for_type_and_chromosome}
\alias{plot_lengths_for_type_and_chromosome}
\title{Plot the distribution of mutation types in the database}
\usage{
plot_lengths_for_type_and_chromosome(endpoint, type, chromosome,
  maximum_length = 5000)
}
\arguments{
\item{endpoint}{The endpoint address to get data from.}

\item{type}{The type to plot for.}

\item{chromosome}{The chromosome to plot for.}

\item{maximum_length}{(optional)  The maximum SV length. [default=5000]}
}
\description{
Plot the distribution of mutation types in the database
}
