% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutation_types.R
\name{plot_mutations_per_type}
\alias{plot_mutations_per_type}
\title{Plot the number of mutations per type in the database.}
\usage{
plot_mutations_per_type(endpoint)
}
\arguments{
\item{endpoint}{The endpoint address to get data from.}
}
\description{
Plot the number of mutations per type in the database.
}
